requirements.txt
loguru==0.7.0
numpy==1.26.4
pyannote.audio==3.3.0
python-dotenv==1.1.1
soundfile==0.13.1
torch==2.8.0
torchaudio==2.8.0
transformers==4.49.0
accelerate>=0.26.0
asteroid>=0.7.0
scikit-learn>=1.3.0
speechbrain>=1.0.0
hyperpyyaml>=1.2.0
ruamel.yaml>=0.18.0
faster-whisper
